fixed schema location for club profile
