+++ web-app/db/schema.ts
@@ -1,4 +1,4 @@
 export * from "./discovery/schema";
 export * from "./account/schema";
 export * from "./attendance/schema";
-export * from "./club-addition/schema";
+export * from "./club-profile/schema";
